refactor(PO): introduce pag object classes inheritance
