+++ src/test/java/LoginFormErrorStateTest.java
@@ -1,4 +1,4 @@
-import LoginPage.LoginPage;
+import pages.LoginPage;
 import io.github.bonigarcia.wdm.WebDriverManager;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
