+++ src/test/java/pages/BasePage.java
@@ -1,0 +1,20 @@
+package pages;
+
+import org.openqa.selenium.WebDriver;
+
+public abstract class BasePage {
+    protected final WebDriver driver;
+    protected final String url;
+
+    public BasePage(WebDriver driver, String url) {
+        this.driver = driver;
+        this.url = url;
+    }
+
+    public void open() {
+        driver.get(url);
+    }
+
+    public abstract boolean isOpened();
+}
+
